Move frameworks related code to separate modules (#19)
* Move framework support to plugins

* Fix code style

* update docstring

* switch smoke test job to on pull_request event

* rename Parameters to FrameworkParameters

* Fix tests

* working imports

* fix relative imports - part 1

* fix relative imports - part 2

* fix tests

* add known frameworks object to avoid using strings

* reduce number of files in openvino framework executor

* remove extra requirements

* fix lynter

* fix linter

* remove commented code

* rename Singleton.py to lowercase

* merge config_processor.py and test_config_parser.py to one file

* merge parameters and parameter parsers

* fix linter errors

* replace .format to f-string in changed files and replace os.path to pathlib

* fix tests

* fix linter error

* fix review comments

* fix linter again

* Update src/benchmark/frameworks/tensorflow/tensorflow_test.py


* remove cat results.txt if failed
