+++ src/benchmark/frameworks/config_parser/dependent_parameters_parser.py
@@ -1,0 +1,7 @@
+import abc
+
+
+class DependentParametersParser(metaclass=abc.ABCMeta):
+    @abc.abstractmethod
+    def parse_parameters(self, curr_test):
+        pass
