+++ src/benchmark/frameworks/config_parser/test_reporter.py
@@ -1,0 +1,13 @@
+import abc
+
+
+class Test(metaclass=abc.ABCMeta):
+    def __init__(self, model, dataset, indep_parameters, dep_parameters):
+        self.model = model
+        self.dataset = dataset
+        self.indep_parameters = indep_parameters
+        self.dep_parameters = dep_parameters
+
+    @abc.abstractmethod
+    def get_report(self):
+        pass
