+++ src/benchmark/frameworks/known_frameworks.py
@@ -1,0 +1,4 @@
+class KnownFrameworks:
+    openvino_dldt = 'OpenVINO DLDT'
+    tensorflow = 'TensorFlow'
+    caffe = 'Caffe'
