+++ src/benchmark/frameworks/openvino/openvino_parameters_parser.py
@@ -1,0 +1,79 @@
+from pathlib import Path
+
+from ..config_parser.dependent_parameters_parser import DependentParametersParser
+from ..config_parser.framework_parameters_parser import FrameworkParameters
+
+
+class OpenVINOParametersParser(DependentParametersParser):
+    def parse_parameters(self, curr_test):
+        CONFIG_FRAMEWORK_DEPENDENT_TAG = 'FrameworkDependent'
+        CONFIG_FRAMEWORK_DEPENDENT_MODE_TAG = 'Mode'
+        CONFIG_FRAMEWORK_DEPENDENT_EXTENSION_TAG = 'Extension'
+        CONFIG_FRAMEWORK_DEPENDENT_ASYNC_REQUEST_COUNT_TAG = 'AsyncRequestCount'
+        CONFIG_FRAMEWORK_DEPENDENT_THREAD_COUNT_TAG = 'ThreadCount'
+        CONFIG_FRAMEWORK_DEPENDENT_STREAM_COUNT_TAG = 'StreamCount'
+
+        dep_parameters_tag = curr_test.getElementsByTagName(CONFIG_FRAMEWORK_DEPENDENT_TAG)[0]
+
+        _mode = dep_parameters_tag.getElementsByTagName(
+            CONFIG_FRAMEWORK_DEPENDENT_MODE_TAG)[0].firstChild
+        _extension = dep_parameters_tag.getElementsByTagName(
+            CONFIG_FRAMEWORK_DEPENDENT_EXTENSION_TAG)[0].firstChild
+        _async_request_count = dep_parameters_tag.getElementsByTagName(
+            CONFIG_FRAMEWORK_DEPENDENT_ASYNC_REQUEST_COUNT_TAG)[0].firstChild
+        _thread_count = dep_parameters_tag.getElementsByTagName(
+            CONFIG_FRAMEWORK_DEPENDENT_THREAD_COUNT_TAG)[0].firstChild
+        _stream_count = dep_parameters_tag.getElementsByTagName(
+            CONFIG_FRAMEWORK_DEPENDENT_STREAM_COUNT_TAG)[0].firstChild
+
+        return OpenVINOParameters(
+            mode=_mode.data if _mode else None,
+            extension=_extension.data if _extension else None,
+            async_request_count=_async_request_count.data if _async_request_count else None,
+            thread_count=_thread_count.data if _thread_count else None,
+            stream_count=_stream_count.data if _stream_count else None,
+        )
+
+
+class OpenVINOParameters(FrameworkParameters):
+    def __init__(self, mode, extension, async_request_count, thread_count, stream_count):
+        self.mode = None
+        self.extension = None
+        self.async_request = None
+        self.nthreads = None
+        self.nstreams = None
+
+        if self._mode_is_correct(mode):
+            self.mode = mode.title()
+        if self._extension_path_is_correct(extension):
+            self.extension = extension
+        else:
+            raise ValueError('Wrong extension path for device. File not found.')
+        if self.mode == 'Sync':
+            if self._parameter_not_is_none(thread_count):
+                if self._int_value_is_correct(thread_count):
+                    self.nthreads = int(thread_count)
+                else:
+                    raise ValueError('Thread count can only take values: integer greater than zero.')
+        if self.mode == 'Async':
+            if self._parameter_not_is_none(async_request_count):
+                if self._int_value_is_correct(async_request_count):
+                    self.async_request = async_request_count
+                else:
+                    raise ValueError('Async requiest count can only take values: integer greater than zero.')
+            if self._parameter_not_is_none(stream_count):
+                if self._int_value_is_correct(stream_count):
+                    self.nstreams = stream_count
+                else:
+                    raise ValueError('Stream count can only take values: integer greater than zero.')
+
+    @staticmethod
+    def _mode_is_correct(mode):
+        const_correct_mode = ['sync', 'async', 'ovbenchmark_python_latency', 'ovbenchmark_python_throughput',
+                              'ovbenchmark_cpp_latency', 'ovbenchmark_cpp_throughput']
+        if mode.lower() in const_correct_mode:
+            return True
+        raise ValueError(f'Mode is a required parameter. Mode can only take values: {", ".join(const_correct_mode)}')
+
+    def _extension_path_is_correct(self, extension):
+        return not self._parameter_not_is_none(extension) or Path(extension).exists()
