+++ src/benchmark/frameworks/openvino/openvino_wrapper.py
@@ -1,0 +1,16 @@
+from .openvino_process_factory import create_process
+from .openvino_test import OpenVINOTest
+from ..framework_wrapper import FrameworkWrapper
+from ..known_frameworks import KnownFrameworks
+
+
+class OpenVINOWrapper(FrameworkWrapper):
+    framework_name = KnownFrameworks.openvino_dldt
+
+    @staticmethod
+    def create_process(test, executor, log, cpp_benchmark_path=None):
+        return create_process(test, executor, log, cpp_benchmark_path)
+
+    @staticmethod
+    def create_test(model, dataset, indep_parameters, dep_parameters):
+        return OpenVINOTest(model, dataset, indep_parameters, dep_parameters)
